Replace assumptions by explicit conditional statements
Guard code sections handling input capture events with an ICF_ check.

Remove continue statement: do not rely on the implicit assumption that
ICF_ and OCF_ are mutually exclusive, because this is not guaranteed.
Further patches to better handle this case will follow.

This diff is best viewed with --ignore-space-change (-b).