Rework timerdiff code to prepare for further changes
Calculate the raw timer difference sooner because prospective code for
timer overflow handling will need it.

After the timerdiff calculation, the old prevVal is not used again,
move the statement updating it as well.

In the bitwise OR, move the operand supplying the most significant bits
to the left, and remove unneccesary brackets.

No semantic change.